Resolve "Core migration from Spring Boot 3.0 to 3.1"

Closes #115

See merge request cm/rnd/cytomine/core/core-ce!42

+++ src/main/java/be/cytomine/repository/ontology/ReviewedAnnotationRepository.java
@@ -19,18 +19,14 @@ package be.cytomine.repository.ontology;
 import be.cytomine.domain.image.ImageInstance;
 import be.cytomine.domain.ontology.ReviewedAnnotation;
 import be.cytomine.domain.ontology.Term;
-import be.cytomine.domain.ontology.UserAnnotation;
 import be.cytomine.domain.project.Project;
 import be.cytomine.domain.security.User;
-import be.cytomine.dto.AnnotationLight;
 import be.cytomine.dto.ReviewedAnnotationStatsEntry;
-import be.cytomine.service.UrlApi;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import org.springframework.data.jpa.repository.Query;
 
-import javax.persistence.Tuple;
-import java.math.BigInteger;
+import jakarta.persistence.Tuple;
 import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
@@ -77,9 +73,9 @@ public interface ReviewedAnnotationRepository extends JpaRepository<ReviewedAnno
         List<ReviewedAnnotationStatsEntry> reviewedAnnotationStatsEntries = new ArrayList<>();
         for (Tuple tuple : stats(imageInstance.getId())) {
             reviewedAnnotationStatsEntries.add(new ReviewedAnnotationStatsEntry(
-                    ((BigInteger)tuple.get(0)).longValue(),
-                    ((BigInteger)tuple.get(1)).longValue(),
-                    ((BigInteger)tuple.get(2)).longValue())
+                    (Long)tuple.get(0),
+                    (Long)tuple.get(1),
+                    (Long)tuple.get(2))
             );
         }
         return reviewedAnnotationStatsEntries;
