+++ src/main/java/be/cytomine/service/ontology/GenericAnnotationService.java
@@ -26,16 +26,13 @@ import be.cytomine.repository.ontology.ReviewedAnnotationRepository;
 import be.cytomine.repository.ontology.UserAnnotationRepository;
 import be.cytomine.service.CurrentUserService;
 import be.cytomine.service.security.SecurityACLService;
-import be.cytomine.service.utils.KmeansGeometryService;
-import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import javax.persistence.EntityManager;
-import javax.persistence.Tuple;
-import javax.transaction.Transactional;
-import java.math.BigInteger;
+import jakarta.persistence.EntityManager;
+import jakarta.persistence.Tuple;
+import jakarta.transaction.Transactional;
 import java.util.*;
 import java.util.stream.Collectors;
 
@@ -91,8 +88,8 @@ public class GenericAnnotationService {
         List<Long> ids = new ArrayList<>();
         Set<Long> users = new HashSet<>();
         for (Tuple result : results) {
-            ids.add(((BigInteger)result.get("annotation")).longValue());
-            users.add(((BigInteger)result.get("user")).longValue());
+            ids.add((Long)result.get("annotation"));
+            users.add((Long)result.get("user"));
         }
 
         if(users.size()>1 && !table.equals("reviewed_annotation")) { //if more annotation from more than 1 user NOT IN REVIEW MODE!
